Show Advisory Speed Limit by default on link presets in Canada, U.S.
Canada and the United States generally post only an advisory speed limit and avoid posting a legal speed limit on highway ramps. Other countries differ. This change splits the maxspeed field into two as a workaround for the inability to limit the default inclusion of a field to a particular location set.