Don't suggest capacity for amenity=parking_space
With `amenity=parking_space` it is possible to map separate parking
spaces in a parking area. It can be used with `capacity` in case the
space is used for more than one vehicle, but this is an exception. By
default, `capacity=1` is implied, and many mappers omit the capacity tag
when its value would be `1`, only using it when it deviates from the
default.

By suggesting adding `capacity=1` to parking spaces lacking the capacity
tag, the editor stimulates adding a tag which is not needed in the vast
majority of the cases, and novice users make edits containing many
of these changes where they are not needed. I find that this sugggestion
is causing more noise than useful contributions.