Added a preset for amenity=ticket_validator (#166)
* Added a preset for amenity=ticket_validator

amenity=ticket_validator (https://wiki.openstreetmap.org/wiki/Tag%3Aamenity%3Dticket_validator) has increased by over 2300 uses by this time, but it din't have a preset yet. I did not open an issue for that, but I can if necessary. So as the usage for this tag seems very clear to me, I thought it would be worth adding preset for it. I used the icon from vending_machine, maybe when you're not okay with it, you can remove it. I hope everything fits.

* Added some terms to amenity=ticket_validator preset

* added aliases to amenity=ticket_validator preset