hotfix for wrongly applied tagging upgrade for parcel lockers
The upgrade paths for the old `vending=parcel_pickup` and `vending=parcel_mail_in` tags overlap for the case of a feature tagged as `parcel_pickup;parcel_mail_in`. This casuses all three matching `replacement`s to be applied in iD. The value of the `parcel_mail_in` conflicts in this case. Since for `parcel_pickup` the implied `yes` default should not be set, it is impossible to resolve this ambiguity with the current schema of `deprecated.json`.
This hotfix/workaround solves this by skipping the upgrade path for `vending=parcel_mail_in`-only features. This minimally impacts the functionality, as such parcel lockers are much less common than pickup-only or combined ones (250 vs. 18k features).

closes #364