Remove "line" geometry from lift gate
Lift gates should only be nodes ([wiki](https://wiki.openstreetmap.org/wiki/Tag%3Abarrier%3Dlift_gate)). This line probably is the cause of a validation issue, where ID suggests converting an orphaned lift_gate node into a way instead of suggesting it to be added as a node to a way.