remove "composite" value
this one is not listed on the wiki page, and is unclear what exactly it is meant to represent:

Composite materials (https://en.wikipedia.org/wiki/Composite_material) can be everything from reinforced concrete, plywood, carbon or glass fiber materials for example. Or it could mean to represent an object made of nor than a single material (e.g. a bench with wooden seating surface and metal structural elements). Until it is clarified what the tag is actually meant to represent, I think it's better to not promote this tag value.