hide "street_cabinet" field; add presets for special cabinets
after #777, unfortunately, there were two fields to define the sub-type of street cabinets: utility for the "common" ones (e.g. telecom, power, etc.) and the "legacy" street_cabinet field for the rest. This PR adds dedicated presets for the legacy cases where the "street_cabinet" tag is still to be used to solve this UI issue.