tweak name
"Townhouse" seems to be the more generic term used in American English according to wikipedia (while "row house" is apparently more regional). This further allows to remove the clarification in brackets by using "row of townhouses" as the label for the building=terrace preset.

Also keep "Row House" as an alias and add variations as search terms.