tweak some strings
mostly to make them a bit clearer. also removed `bottle_return` value: this one is special, as it is not actually a "good" that is sold at the vending machine (it is merely used because there is no better main tag for such machines that afford returning of used bottles). It is unlikely that a mapper selects the Vending Machine preset to map a bottle-return station, instead of just searching for the dedicated preset directly.