don't always show "wikipedia" field when the wikidata tag is set
as not all wikidata items have a matching wikipedia page, this field would inevitably be always empty in these cases. As it is automatically filled in when a user selects a wikidata entry with a matching wikipedia page, the field shows up in the "regular" case anyway. only if the wikipedia page exists, but a map feature only has the wikidata tag (e.g. because it was mapped by another editor, or added manually in the raw tag editor), the field is now missing; it can still be added manually through "more fields" in this case, though

see https://github.com/openstreetmap/id-tagging-schema/pull/964#pullrequestreview-1541084558