Introduce `cycleway:both=*` and make `cycleway:NULL=*` unsearchable
The goal is, to work around
- a deprecation of the `:NULL` variant that used to represent "both" or sometimes "the primary side"
- the need to add another feature to the files which would allow to handle both keys at the same time.

The idea is, to have the legacy value kind of unsearchable and only show up when the `cycleway:NULL` key is present. However, that is hard ATM because this key is also used for `crossing`, `asl` and such (https://taginfo.openstreetmap.org/keys/cycleway#values) and the `prerequisiteTag` does not allow any arrays as values, yet.