crossing: New approach with `@templates/crossing/bicycle_more`, adding `segregated`
The fields `oneway` and `access` are important for `highway=cycleway|path` crossings but not essential. They are more of a advanced user setup which should be visible when prev filled in but only added by users that read more about it before. They are moved to the `moreFields` for that reason.

The `segregated` is added here for the same reasons and because of it's importance for highway types that likely have bike traffic.
Ping https://github.com/openstreetmap/id-tagging-schema/issues/317

The `surface` and `smoothness` is extracted from the `@template` because it makes more sense to split them up in `fields` and `moreFields`. A templates adds too much abstraction in this case.