Accept units in Elevation field
The Elevation field is now of type roadheight, so that the user can explicitly specify a unit other than meters when appropriate without having to convert the quantity in their head. Technically, this is not a road height but rather a different height dimension, but it happens to be measured in the same units in both metric and customary countries.