Add deprecation rules for cycleway=opposite: remove oneway=yes
We found cases where oneway=yes is not the right update path because access tags are used instead. We assume, that oneway=yes is usually tagged already, so we don't need to add it to the replacement. However, the oneway:bicycle=no is used less often and also does less harm in case someone does apply the deprecation in cases where it should not be applied.