explicitly ban AI slop
Though I have doubts about whether anyone making this LLM pull requests will read it.

It intentionally makes note that human-written content with AI-slop characteristic is bannable. Not because I consider it likely, but because if someone starts claiming that their AI slop is not AI slop, it is not necessary to prove how they obtained it.